fix: archive arion build on macos
streamLayeredImage will generate a script with shebangs for the system
used to generate the actual image, which means it can't be run on a host
with a different system. this is a the situation of an M1 macOS using
linux-builder. the workaround is to force buildLayeredImage (will be
fully built on the linux-builder and passed to docker-compose directly)
when the image system does not match the host system.