fix: use hashicorp/<provider> for original builtin
it will be fetched from the opentofu registry using opentofu prefix if
it was hashicorp but otherwise treated the same. the reason this is
necessary is when using external modules, they will default to the name
hashicorp/<provider>, and we can't have a provider naming conflict. best
to avoid duplicates essentially and just use the original name